DEV9: Move helper Qt classes into a separate file
