+++ pcsx2-qt/CMakeLists.txt
@@ -93,6 +93,8 @@ target_sources(pcsx2-qt PRIVATE
 	Settings/DEV9SettingsWidget.cpp
 	Settings/DEV9SettingsWidget.h
 	Settings/DEV9SettingsWidget.ui
+	Settings/DEV9UiCommon.cpp
+	Settings/DEV9UiCommon.h
 	Settings/HddCreateQt.cpp
 	Settings/HddCreateQt.h
 	Settings/SettingsDialog.cpp
