+++ pcsx2-qt/Settings/DEV9SettingsWidget.h
@@ -16,7 +16,6 @@
 #pragma once
 
 #include <QtWidgets/QWidget>
-#include <QtWidgets/QItemDelegate>
 #include <QtGui/QStandardItemModel>
 
 #include "ui_DEV9SettingsWidget.h"
@@ -24,35 +23,6 @@
 #include "DEV9/net.h"
 
 class SettingsDialog;
-
-class IPValidator : public QValidator
-{
-	Q_OBJECT
-
-public:
-	explicit IPValidator(QObject* parent = nullptr, bool allowEmpty = false);
-	virtual State validate(QString& input, int& pos) const override;
-
-private:
-	static const QRegularExpression intermediateRegex;
-	static const QRegularExpression finalRegex;
-
-	bool m_allowEmpty;
-};
-
-class IPItemDelegate : public QItemDelegate
-{
-	Q_OBJECT
-
-public:
-	explicit IPItemDelegate(QObject* parent = nullptr);
-
-protected:
-	QWidget* createEditor(QWidget* parent, const QStyleOptionViewItem& option, const QModelIndex& index) const;
-	void setEditorData(QWidget* editor, const QModelIndex& index) const;
-	void setModelData(QWidget* editor, QAbstractItemModel* model, const QModelIndex& index) const;
-	void updateEditorGeometry(QWidget* editor, const QStyleOptionViewItem& option, const QModelIndex& index) const;
-};
 
 class DEV9SettingsWidget : public QWidget
 {
@@ -86,13 +56,6 @@ protected:
 	bool eventFilter(QObject* object, QEvent* event);
 
 private:
-	struct HostEntryUi
-	{
-		std::string Url;
-		std::string Desc;
-		std::string Address = "0.0.0.0";
-		bool Enabled;
-	};
 
 	void AddAdapter(const AdapterEntry& adapter);
 	void RefreshHostList();
